Move for loop & if block reconstructors to their own files
