+++ lib/utils/templates/reconstruct.util.js
@@ -1,6 +1,9 @@
 const { NODE_TYPES } = require('../../constants/ast');
 const { checkIsType, TYPES } = require('../types.util');
 
+const reconstructForLoop = require('./reconstructors/for.reconstruct');
+const reconstructIfBlock = require('./reconstructors/if.reconstruct');
+
 const reconstruct = (tree, metadata, contents) => {
   let content = '';
 
@@ -10,25 +13,9 @@ const reconstruct = (tree, metadata, contents) => {
   } else if (tree.type === NODE_TYPES.PARTIAL) {
     content += reconstruct(tree.content, metadata);
   } else if (tree.type === NODE_TYPES.FOR_LOOP) {
-    const { variable } = tree.content;
-    const metadataValue =
-      metadata[variable] ||
-      (checkIsType(contents, TYPES.OBJECT) && contents[variable]);
-
-    if (Array.isArray(metadataValue)) {
-      for (const value of metadataValue) {
-        content += reconstruct(tree.content.child, metadata, value);
-      }
-    }
+    content += reconstructForLoop(tree, metadata, contents, reconstruct);
   } else if (tree.type === NODE_TYPES.IF_BLOCK) {
-    const { child, condition } = tree.content;
-    const match = condition.match(/(.+)(===|!==|>=|>|<=|<)(.+)/s);
-
-    if ((!match || match.length === 1) && metadata[condition]) {
-      content += reconstruct(child, metadata);
-    } else if (match && checkConditionTrue(match, metadata)) {
-      content += reconstruct(child, metadata);
-    }
+    content += reconstructIfBlock(tree, metadata, reconstruct);
   }
 
   content = replaceVariables(content, metadata, contents);
@@ -39,66 +26,6 @@ const reconstruct = (tree, metadata, contents) => {
 
   return content;
 };
-
-const checkConditionTrue = (match, metadata) => {
-  const lhs = match[1].trim();
-  const op = match[2].trim();
-  const rhs = match[3].trim();
-  let isTrue = false;
-  let eqValue;
-  let value;
-
-  /* istanbul ignore else */
-  if (lhs in metadata) {
-    value = metadata[lhs];
-    eqValue = rhs;
-  } else if (rhs in metadata) {
-    value = metadata[rhs];
-    eqValue = lhs;
-  }
-
-  /* istanbul ignore else */
-  if (value && eqValue) {
-    isTrue = evaluateCondition(op, value, eqValue);
-  }
-
-  return isTrue;
-};
-
-const evaluateCondition = (op, value, eqValue) => {
-  const evalValue = convertValue(value);
-  const evalEqValue = convertValue(eqValue);
-
-  switch (op) {
-    case '===':
-      return evalValue === evalEqValue;
-    case '!==':
-      return evalValue !== evalEqValue;
-    case '<':
-      return evalValue < evalEqValue;
-    case '>=':
-      return evalValue >= evalEqValue;
-    case '>':
-      return evalValue > evalEqValue;
-    case '<=':
-      return evalValue <= evalEqValue;
-    /* istanbul ignore next */
-    default:
-      return false;
-  }
-};
-
-const convertValue = (value) => {
-  let converted = value;
-
-  try {
-    converted = eval(value);
-  } catch (e) {
-    converted = value;
-  }
-
-  return converted;
-}
 
 const replaceVariables = (content, metadata, contents) => {
   const variableRegex = /<!--\s*talc:([a-zA-Z0-9_]+)\s*-->/;
